Create docker-compose.yaml
How to setup Passbolt with Synology.

MariaDB with docker is either working or not. Whenever I tested the permissions were a mess, and it either created the user without the permissions or not at all.

This setup suggests that you setup a separate MariaDB already in your environment, with the relevant accesses.